added learner group model test file
